Move indexing into query file
